Move load from bigchain-benchmark to bigchain
