+++ bigchaindb/commands/bigchain.py
@@ -11,8 +11,13 @@ import copy
 import json
 import builtins
 
+import logstats
+
+
 import bigchaindb
 import bigchaindb.config_utils
+from bigchaindb.util import ProcessGroup
+from bigchaindb.client import temp_client
 from bigchaindb import db
 from bigchaindb.exceptions import DatabaseAlreadyExists, KeypairNotFoundException
 from bigchaindb.commands import utils
@@ -160,6 +165,38 @@ def run_start(args):
     processes.start()
 
 
+def _run_load(tx_left, stats):
+    logstats.thread.start(stats)
+    client = temp_client()
+    # b = bigchaindb.Bigchain()
+
+    while True:
+        tx = client.create()
+
+        stats['transactions'] += 1
+
+        if tx_left is not None:
+            tx_left -= 1
+            if tx_left == 0:
+                break
+
+
+def run_load(args):
+    bigchaindb.config_utils.autoconfigure(filename=args.config, force=True)
+    logger.info('Starting %s processes', args.multiprocess)
+    stats = logstats.Logstats()
+    logstats.thread.start(stats)
+
+    tx_left = None
+    if args.count > 0:
+        tx_left = int(args.count / args.multiprocess)
+
+    workers = ProcessGroup(concurrency=args.multiprocess,
+                           target=_run_load,
+                           args=(tx_left, stats.get_child()))
+    workers.start()
+
+
 def main():
     parser = argparse.ArgumentParser(
         description='Control your BigchainDB node.',
@@ -199,6 +236,24 @@ def main():
     # parser for starting BigchainDB
     subparsers.add_parser('start',
                           help='Start BigchainDB')
+
+    load_parser = subparsers.add_parser('load',
+                                        help='Write transactions to the backlog')
+
+    load_parser.add_argument('-m', '--multiprocess',
+                             nargs='?',
+                             type=int,
+                             default=False,
+                             help='Spawn multiple processes to run the command, '
+                                  'if no value is provided, the number of processes '
+                                  'is equal to the number of cores of the host machine')
+
+    load_parser.add_argument('-c', '--count',
+                             default=0,
+                             type=int,
+                             help='Number of transactions to push. If the parameter -m '
+                                  'is set, the count is distributed equally to all the '
+                                  'processes')
 
     utils.start(parser, globals())
